pyproject.toml
[project]
name = "memcore"
version = "0.1.0"
description = "Standalone agentic memory management system for AI agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "fastembed>=0.6.0",
    "jinja2>=3.1.0",
    "litellm>=1.81.14",
    "mcp>=1.26.0",
    "networkx>=3.6.1",
    "onnxruntime>=1.20.0",
    "psutil>=5.9.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.0.0",
    "qdrant-client>=1.17.0",
    "starlette>=0.45.0",
    "strands-sdk>=10.0.2",
    "uvicorn>=0.34.0",
    "watchdog>=6.0.0",
]

[project.optional-dependencies]
tray = ["pystray>=0.19.5", "pillow>=11.0.0", "requests>=2.32.0"]
all = ["pystray>=0.19.5", "pillow>=11.0.0", "requests>=2.32.0"]

[project.scripts]
memcore-tray = "src.memcore.tray.app:run_tray_app"
